requirements.txt
langdetect==1.0.9
numpy==1.23.4
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
pycryptodome==3.15.0
pyparsing==3.0.9
PyPDF2==2.11.1
pytesseract==0.3.10
six==1.16.0
